+++ requirements/requirements.txt
@@ -1,5 +1,2 @@
-# bs4
-# Faker
-# lxml
 requests
 pyrate-limiter
